Mining News Scrapper/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.30.0
python-dotenv>=1.0.0
